Home and Weather page layouts/display @MVP+git add . Moving on to Settings and Quote provider
