+++ src/components/Home.css
@@ -9,20 +9,17 @@
     transition: 0.4 ease-out;
 }
 
+.backgroundFilter {
+    background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.75));    
+    min-height: 100vh;
+    padding: 25px;
+}
+
 .main--home {
     display: flex;
-    /* justify-content: center; */
     flex-direction: column;
-    /* align-items: center;
-    align-content: center; */
-    color: black;
-    height: 600px;
-    /* border: 1px solid green; */
-    /* min-height: 100vh; */
-    /* background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.75)); */
-    /* padding: 25px; */
-    
-    
+    justify-content: space-evenly;
+    align-items: center;
 }
 
 .dailyQuote--homeDiv {
@@ -30,46 +27,53 @@
     flex-direction: column;
     align-content: center;
     align-self: flex-start;
-    /* border: 1px solid gold; */
     justify-content: center;
     width: 100%;
     padding: 1em;
 }
 
 .dailyQuote--home {
-    /* border: 1px solid black; */
-    font-size: 12px;
-    padding: 1em;
+    color: snow;
+    font-size: 14px;
+    font-weight: 500;
+    text-align: left;
+    text-shadow: 3px 3px rgba(50, 50, 70, 0.5);
+    padding-left: 1em;
 }
 
 .dailyQuoteSource--home {
-    /* border: 1px solid black; */
-    font-size: 12px;
+    color: snow;
+    font-size: 14px;
+    font-weight: 500;
+    text-shadow: 3px 3px rgba(50, 50, 70, 0.5);
     text-align: right;
-    padding-right: 2em;
+    padding-right: 1em;
+}
+
+.date {
+    color: snow;
+    font-size: 20px;
+    font-weight: 300;
+    font-style: italic;
+    text-align: center;
+    text-shadow: 2px 2px rgba(50, 50, 70, 0.5);
 }
 
 .localTemp--homeDiv {
     justify-content: center;
     align-content: center;
     align-self: center;
-    /* border: 1px solid blue; */
     width: 100%;
     padding: 1em;
-    margin: 2.5em;
 }
 
 .localTemp--home {
-    /* border: 1px solid black; */
-    /* text-align: center;
-    font-size: 8em;
-    color: gray; */
     background-color: rgba(255, 255, 255, 0.2);
     border-radius: 16px;
 
     padding: 15px 25px;
 
-    color: #FFF;
+    color: snow;
     font-size: 100px;
     font-weight: 900;
