+++ src/components/weather/Weather.js
@@ -22,17 +22,17 @@ export const WeatherApp = () => {
         }
     }
 
-    const dateBuilder = (d) => {
-        let months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
-        let days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+    // const dateBuilder = (d) => {
+    //     let months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
+    //     let days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
 
-        let day = days[d.getDay()];
-        let date = d.getDate();
-        let month = months[d.getMonth()];
-        let year = d.getFullYear();
+    //     let day = days[d.getDay()];
+    //     let date = d.getDate();
+    //     let month = months[d.getMonth()];
+    //     let year = d.getFullYear();
 
-        return `${day} ${date} ${month} ${year}`
-    }
+    //     return `${day} ${date} ${month} ${year}`
+    // }
 
 
     return (
@@ -53,7 +53,7 @@ export const WeatherApp = () => {
                     <div>
                         <div className="location-box">
                             <div className="location">{weather.name}, {weather.sys.country}</div>
-                            <div className="date">{dateBuilder(new Date())}</div>
+                            {/* <div className="date">{dateBuilder(new Date())}</div> */}
                         </div>
                         <div className="weather-box">
                             <div className="temp">
